+++ app/src/main/java/com/plex/androidsdk/httpdatasources/IBaseInput.java
@@ -21,7 +21,7 @@
 package com.plex.androidsdk.httpdatasources;
 
 /**
- * Used to identify classes that contain the input parameters for a request.
+ * Marker interface used to identify classes that contain the input parameters for a request.
  */
 public interface IBaseInput {