+++ app/src/main/java/com/plex/androidsdk/httpdatasources/Inventory/Container_Get1.java
@@ -20,13 +20,14 @@
 
 package com.plex.androidsdk.httpdatasources.Inventory;
 
-import com.google.gson.JsonArray;
 import com.google.gson.annotations.SerializedName;
 import com.plex.androidsdk.httpdatasources.BaseRow;
 import com.plex.androidsdk.httpdatasources.DataSource;
 import com.plex.androidsdk.httpdatasources.HttpDataSourceCredentials;
 import com.plex.androidsdk.httpdatasources.IBaseInput;
 import com.plex.androidsdk.httpdatasources.IDataSourceCallback;
+import com.plex.androidsdk.httpdatasources.IDataSourceConnector;
+import java.lang.reflect.Type;
 import java.math.BigDecimal;
 
 /**
@@ -51,6 +52,13 @@ public class Container_Get1 extends DataSource {
    */
   public Container_Get1(IDataSourceCallback callback, HttpDataSourceCredentials credentials, String serverName, boolean test) {
     super(callback, credentials, serverName, test);
+  }
+
+  /**
+   * {@inheritDoc}
+   */
+  protected Container_Get1(IDataSourceCallback callback, HttpDataSourceCredentials credentials, String serverName, boolean test, IDataSourceConnector connector) {
+    super(callback, credentials, serverName, test, connector);
   }
 
   /**
@@ -83,7 +91,6 @@ public class Container_Get1 extends DataSource {
     return 6455;
   }
 
-
   /**
    * The input parameters to be used for the data source call. {@inheritDoc}
    *
@@ -94,75 +101,22 @@ public class Container_Get1 extends DataSource {
     return inputParameters;
   }
 
-  /**
-   * Parse the JSON for a returned row and assign to the Row class.
-   *
-   * @param rowArray A row entry in the returned JSON.
-   * @return Row Contains the values from the parsed JSON. Returns null if no data.
-   */
   @Override
-  protected BaseRow parseRow(JsonArray rowArray) {
-    Row row = null;
-
-    // Make sure there is data
-    if (rowArray.size() > 0) {
-      row = new Row();
-
-      // Loop through all the fields and assign the data to Row.
-      for (int i = 0; i < rowArray.size(); ++i) {
-        switch (i) {
-          case 0: // Part No Revision
-            row.setPartNoRevision(rowArray.get(0).getAsString());
-            break;
-          case 1: // Name
-            row.setName(rowArray.get(1).getAsString());
-            break;
-          case 2: // Part Key
-            row.setPartKey(rowArray.get(2).getAsInt());
-            break;
-          case 3: // Operation Code
-            row.setOperationCode(rowArray.get(3).getAsString());
-            break;
-          case 4: // Quantity
-            row.setQuantity(rowArray.get(4).getAsBigDecimal());
-            break;
-          case 5: // Container Status
-            row.setContainerStatus(rowArray.get(5).getAsString());
-            break;
-          case 6: // Location
-            row.setLocation(rowArray.get(6).getAsString());
-            break;
-          case 7: // Note
-            row.setNote(rowArray.get(7).getAsString());
-            break;
-          case 8: // Operation Key
-            if (rowArray.get(8).isJsonNull() == false) {
-              row.setOperationKey(rowArray.get(8).getAsInt());
-            }
-            break;
-          case 9: // Rework Operation
-            if (rowArray.get(9).isJsonNull() == false) {
-              row.setReworkOperation(rowArray.get(9).getAsInt());
-            }
-            break;
-          case 10: // Special Instructions
-            row.setSpecialInstructions(rowArray.get(10).getAsString());
-            break;
-          case 11: // Defect Type
-            row.setDefectType(rowArray.get(11).getAsString());
-            break;
-        }
-      }
-    }
-
-    return row;
+  protected Type getOutputType() {
+    return null;
   }
+
+  @Override
+  protected Type getRowType() {
+    return Row.class;
+  }
+
   //endregion
 
   //region INTERNAL CLASSES
 
   /**
-   * Input parameters for data source call. Used by GSON to serialize into JSON.
+   * Input parameters for data source call. Used by Gson to serialize into JSON.
    */
   private class InputParameters implements IBaseInput {
 
@@ -183,17 +137,29 @@ public class Container_Get1 extends DataSource {
    */
   public class Row extends BaseRow {
 
+    @SerializedName("Container_Status")
     private String containerStatus;
+    @SerializedName("Defect_Type")
     private String defectType;
+    @SerializedName("Location")
     private String location;
+    @SerializedName("Name")
     private String name;
+    @SerializedName("Note")
     private String note;
+    @SerializedName("Operation_Code")
     private String operationCode;
+    @SerializedName("Operation_Key")
     private int operationKey;
+    @SerializedName("Part_Key")
     private int partKey;
+    @SerializedName("Part_No_Revision")
     private String partNoRevision;
+    @SerializedName("Quantity")
     private BigDecimal quantity;
+    @SerializedName("Rework_Operation")
     private int reworkOperation;
+    @SerializedName("Special_Instructions")
     private String specialInstructions;
 
     public String getSpecialInstructions() {